+++ poptorch/include/poptorch/ImplicitCasting.hpp
@@ -20,10 +20,9 @@ enum class HalfFloatCasting { FloatDowncastToHalf = 0, HalfUpcastToFloat };
 
 void setHalfFloatCastingBehavior(HalfFloatCasting behavior);
 void setRunningVarianceAlwaysFloat(bool);
+HalfFloatCasting halfFloatCastingBehavior();
 
 bool runningVarianceAlwaysFloat();
-
-HalfFloatCasting halfFloatCastingBehavior();
 
 enum class ImplicitCast {
   None,