+++ poptorch/source/PoptorchSymbols.cpp
@@ -111,6 +111,10 @@ c10::Symbol begin_multi_conv;
 c10::Symbol multi_conv_part;
 c10::Symbol end_multi_conv;
 c10::Symbol host_side_cast;
+c10::Symbol begin_autocast;
+c10::Symbol suppress_autocast;
+c10::Symbol restore_autocast;
+c10::Symbol autocast;
 
 c10::Symbol end_if;
 c10::Symbol start_if_true;
@@ -135,7 +139,11 @@ static void initializePoptorchSymbols() {
   SYMBOL_INIT(poptorch, begin_multi_conv)
   SYMBOL_INIT(poptorch, multi_conv_part)
   SYMBOL_INIT(poptorch, end_multi_conv)
+  SYMBOL_INIT(poptorch, begin_autocast)
+  SYMBOL_INIT(poptorch, suppress_autocast)
+  SYMBOL_INIT(poptorch, restore_autocast)
   SYMBOL_INIT(poptorch, host_side_cast)
+  SYMBOL_INIT(poptorch, autocast)
 
   SYMBOL_INIT(poptorch, end_if);
   SYMBOL_INIT(poptorch, start_if_true);